fallback to default metrics is URL is invalid
